+++ cmd/operator-sdk/alpha/cmd.go
@@ -16,6 +16,7 @@ package alpha
 
 import (
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/alpha/kubebuilder"
+	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/alpha/scorecard"
 
 	"github.com/spf13/cobra"
 )
@@ -28,6 +29,7 @@ func NewCmd() *cobra.Command {
 
 	cmd.AddCommand(
 		kubebuilder.NewCmd(),
+		scorecard.NewCmd(),
 	)
 	return cmd
 }
